Refactor server routes
